[Core] Introduce MultiGPUExecutor abstract class
As a superclass of RayGPUExecutor, in preparation for introducing a sibling MultiprocessingGPUExecutor
